% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.corpus.textmeta.R
\name{as.corpus.textmeta}
\alias{as.corpus.textmeta}
\title{Transform textmeta to corpus}
\usage{
as.corpus.textmeta(object, docnames = "id",
  docvars = setdiff(colnames(object$meta), "id"),
  metadoc = character(), ...)
}
\arguments{
\item{object}{\code{\link{textmeta}} object}

\item{docnames}{Character: string with the column of object$meta
which should be kept as \code{\link[quanteda]{docnames}}.}

\item{docvars}{Character: vector with columns of object$meta which
should be kept as \code{\link[quanteda]{docvars}}.}

\item{metadoc}{Character: vector with columns of object$meta which
should be kept as \code{\link[quanteda]{metadoc}}.}

\item{...}{Additional parameters like \code{compress} for \code{\link[quanteda]{corpus}}.}
}
\value{
\code{\link[quanteda]{corpus}} object
}
\description{
Transfers data from a \code{\link{textmeta}} object to a
\code{\link[quanteda]{corpus}} object -  the way text data is stored in the
package \code{\link[quanteda]{quanteda}}.
}
\keyword{manip}
