% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.textmeta.corpus.R
\name{as.textmeta.corpus}
\alias{as.textmeta.corpus}
\title{Transform corpus to textmeta}
\usage{
as.textmeta.corpus(corpus, cols, dateFormat = "\%Y-\%m-\%d",
  idCol = "id", dateCol = "date", titleCol = "title",
  textCol = "texts", duplicateAction = TRUE, addMetadata = TRUE)
}
\arguments{
\item{corpus}{object of class \code{\link[quanteda]{corpus}},
package \code{\link[quanteda]{quanteda}}}

\item{cols}{\code{character} vector with columns which should be kept}

\item{dateFormat}{\code{character} string with the date format in the
date column for \code{\link{as.Date}}}

\item{idCol}{\code{character} string with column name of the IDs in corpus
- named "id" in the resulting data.frame}

\item{dateCol}{\code{character} string with column name of the Dates in corpus
- named "date" in the resulting data.frame}

\item{titleCol}{\code{character} string with column name of the Titles in corpus
- named "title" in the resulting data.frame}

\item{textCol}{\code{character} string with column name of the Texts in corpus
- results in a named list ("id") of the Texts}

\item{duplicateAction}{\code{logical}
whether \code{\link{deleteAndRenameDuplicates}} should be applied to the
created \code{\link{textmeta}} object}

\item{addMetadata}{\code{logical} whether the metadata flag of corpus should
be added to the meta flag of the \code{\link{textmeta}} object - if there are
conflicts regarding the naming of columns, the metadata columns would be
overwritten by the document specific columns}
}
\value{
\code{\link{textmeta}} object
}
\description{
Transfers data from a \code{\link[quanteda]{corpus}} object -  the way text
data is stored in the package \code{\link[quanteda]{quanteda}} - to a
\code{\link{textmeta}} object.
}
\keyword{manip}
