% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteAndRenameDuplicates.R
\name{deleteAndRenameDuplicates}
\alias{deleteAndRenameDuplicates}
\title{Delete And Rename Articles with the same ID}
\usage{
deleteAndRenameDuplicates(object, paragraph = FALSE)
}
\arguments{
\item{object}{A textmeta-object as a result of a read-function.}

\item{paragraph}{Logical: Should be set to \code{TRUE} if the article is a
list of character strings, representing the paragraphs.}
}
\value{
tba
}
\description{
Deletes Articles with the same ID and same text. Renames the ID of Articles
with the same ID but different text-component (_IDFakeDup, _IDRealDup).
}
\details{
Summary: Different types of duplicates:
 "complete duplicates" = same ID, same information in text, same information in meta
 "real duplicates" = same ID, same information in text, different information in meta
 "fake duplicates" = same ID, different information in text
}
