% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docLDA.R
\name{docLDA}
\alias{docLDA}
\title{create lda-ready dataset}
\usage{
docLDA(text, vocab, ldacorrect = TRUE, excludeNA = TRUE, reduce = TRUE)
}
\arguments{
\item{text}{a list of tokenized texts.}

\item{vocab}{a character vector containing all words which should beused for
lda.}

\item{ldacorrect}{logical: If \code{TRUE}, every word repetition gets an own
column.}

\item{excludeNA}{logical: Should \code{NA}s be removed?}

\item{reduce}{logical: Should empty texts be deleted?}
}
\value{
A list in which every entry contains a matrix with two rows: the
first row gives the number of the entry of the word in \code{vocab} minus
one, the second row the number of the occurrence of the word in the article.
If \code{ldacorrect=TRUE} the second row is 1 and the number of the
occurrence of the word will be shown by the number of columns belonging to
this word.
}
\description{
This function transforms a text corpus like the result from
\code{\link{makeClear}} in a form needed by the lda-package.
}
\examples{

##---- Should be DIRECTLY executable !! ----
}
\keyword{manip}
