% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplist.R
\name{duplist}
\alias{duplist}
\title{Creating List of Duplicates}
\usage{
duplist(object, paragraph = FALSE)
}
\arguments{
\item{object}{A textmeta-object.}

\item{paragraph}{Logical: Should be set to \code{TRUE} if the article is a
list of character strings, representing the paragraphs.}
}
\value{
Named List:
\item{uniqueTexts}{ character}
\item{fullyUniqueTexts}{ character}
\item{idFakeDups}{ list}
\item{idRealDups}{ list}
\item{allTextDups}{ list}
\item{textOnlyDups}{ list}
\item{textMetaDups}{ list}
\item{textOthersDups}{ character}
}
\description{
Creates a List of different types of Duplicates in a textmeta-object.
}
\details{
This function helps to identify different types of Duplicates and gives the
ability to exclude these for further Analysis (e.g. LDA).
}
\examples{

}
\keyword{manip}
