\name{holbrook}
\alias{holbrook}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Selected passages of mispelled texts from Holbrook
}
\description{
  The passages in holbrook are taken David Holbrook's 'English for the Rejected' (Cambridge University Press, 1964). They are taken from the the writings of secondary-school children from Holbrook's English class. These excerpts were downloaded from http://www.dcs.bbk.ac.uk/~roger/corpora.html, and names, tags, carriage returns, and punctuation were removed.
}
\usage{
  data(holbrook)
}
\format{
  A vector of 93 character strings.
}
