% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intruderTopics.R
\name{intruderTopics}
\alias{intruderTopics}
\title{Function to validate the fit of the LDA model}
\usage{
intruderTopics(text = NULL, beta = NULL, theta = NULL, id = NULL,
  numIntruder = 1, numOuttopics = 4, byScore = TRUE, minWords = 0L,
  minOuttopics = 0L, stopTopics = NULL, printSolution = FALSE,
  oldResult = NULL, test = FALSE, testinput = NULL)
}
\arguments{
\item{text}{tba}

\item{beta}{A matrix of word-probabilities or frequency table for the topics. Each row is a topic, each column a word. The rows will be divided by the row sums, if they are not 1.}

\item{theta}{tba}

\item{id}{tba}

\item{numIntruder}{Intended number of intruder words. If \code{numIntruder} is a integer vector, the number would be sampled for each topic.}

\item{numOuttopics}{tba Integer: Number of words per topic, including the intruder words}

\item{byScore}{Logical: Should the score of \code{top.topic.words} from the \code{lda} package be used?}

\item{minWords}{tba}

\item{minOuttopics}{tba}

\item{stopTopics}{tba}

\item{printSolution}{tba}

\item{oldResult}{Result object from an unfinished run of \code{intruderWords}. If oldResult is used, all other parameter will be ignored.}

\item{test}{Logical: Enables test mode}

\item{testinput}{Input for function tests}
}
\value{
tba
}
\description{
This function validates a LDA result by presenting a mix of topics and intruder topics to a human user, who has to identity them.
}
\references{
tba
}
\seealso{
tba
}
