% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intruderWords.R
\name{intruderWords}
\alias{intruderWords}
\title{Function to validate the fit of the LDA model}
\usage{
intruderWords(beta = NULL, byScore = TRUE, numTopwords = 30L,
  numIntruder = 1L, numOutwords = 5L, noTopic = TRUE,
  printSolution = FALSE, oldResult = NULL, test = FALSE,
  testinput = NULL)
}
\arguments{
\item{beta}{A matrix of word-probabilities or frequency table for the topics. Each row is a topic, each column a word. The rows will be divided by the row sums, if they are not 1.}

\item{byScore}{Logical: Should the score of \code{top.topic.words} from the \code{lda} package be used?}

\item{numTopwords}{The number of topwords to be used for the intruder words}

\item{numIntruder}{Intended number of intruder words. If \code{numIntruder} is a integer vector, the number would be sampled for each topic.}

\item{numOutwords}{Integer: Number of words per topic, including the intruder words.}

\item{noTopic}{Logical: Is \code{x} input allowed to mark nonsense topics?}

\item{printSolution}{tba}

\item{oldResult}{Result object from an unfinished run of \code{intruderWords}. If oldResult is used, all other parameter will be ignored.}

\item{test}{Logical: Enables test mode}

\item{testinput}{Input for function tests}
}
\value{
tba
}
\description{
This function validates a LDA result by presenting a mix of words from a topic and intruder words to a human user, who has to identity them.
}
\references{
tba
}
\seealso{
tba
}
