% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeClear.R
\name{makeClear}
\alias{makeClear}
\title{Some data preprocessing}
\usage{
makeClear(object, text, sw = "en", paragraph = FALSE)
}
\arguments{
\item{object}{\code{\link{textmeta}} object}

\item{text}{not necassary if \code{object} is specified, else should be
\code{object\$text}: list of article texts}

\item{sw}{\code{character} vector of stopwords. If the vector is of length
one, \code{sw} is interpretated as argument for \code{\link{stopwords}}}

\item{paragraph}{Logical: Should be set to \code{TRUE} if the article is a
list of character strings, representing the paragraphs.}
}
\value{
A list containing the preprocessed Article.
}
\description{
Removes punctuation, numbers and stopwords, change into lowercase letters
and tokenization.
}
\details{
Removes punctuation, numbers and stopwords, change into lowercase letters
and tokenization. Additional some cleaning steps: remove empty words /
paragraphs / article.
}
\examples{

##---- Should be DIRECTLY executable !! ----
}
\keyword{manip}
