% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_curve.R
\name{plot_curve}
\alias{plot_curve}
\title{plot_curve}
\usage{
plot_curve(v = NULL, norm = "tombaugh", test = NULL)
}
\arguments{
\item{v}{this argument is automatically specified by the
Shiny app}

\item{norm}{the normative data. currently, only \code{"tombaugh"}
is implemented}

\item{test}{character. This argument is specified by the
shiny app automatically}
}
\description{
draws the TMT plot in the shiny app
}
