% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Create ggplot histogram}
\usage{
plot_histogram(df, x, title)
}
\arguments{
\item{df}{Dataframe}

\item{x}{Variable to use when creating histogram}

\item{title}{Title for the plot}
}
\description{
Create ggplot histogram
}
