% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScot.R
\name{plotScot}
\alias{plotScot}
\title{Plotting Counts of Documents or Words over Time (relative to Corpus)}
\usage{
plotScot(object, id = object$meta$id, type = c("docs", "words"),
  rel = FALSE, mark = TRUE, unit = "month", curves = c("exact",
  "smooth", "both"), smooth = 0.05, main, xlab, ylim, both.lwd, both.col,
  both.lty, natozero = TRUE, file, ...)
}
\arguments{
\item{object}{\code{\link{textmeta}} object - with strictly tokenized
\code{text} component (\code{character} vectors) if \code{type = "words"}}

\item{id}{\code{character} vector (default: \code{object$meta$id}) which IDs
specify the subcorpus}

\item{type}{\code{character} (default: \code{"docs"}) should counts/proportion
of documents \code{"docs"} or words \code{"words"} be plotted}

\item{rel}{\code{logical} (default: \code{FALSE}) should counts
(\code{FALSE}) or proportion (\code{TRUE}) be plotted}

\item{mark}{\code{logical} (default: \code{TRUE}) should years be marked by
vertical lines}

\item{unit}{\code{character} (default: \code{"month"}) to which unit should
dates be floored. Other possible units are \code{"bimonth}, \code{"quarter}, \code{"season"},
\code{"halfyear"}, \code{"year"}, for more units see \code{\link[lubridate]{round_date}}}

\item{curves}{\code{character} (default: \code{"exact"}) should \code{"exact"},
\code{"smooth"} curve or \code{"both"} be plotted}

\item{smooth}{\code{numeric} (default: \code{0.05}) smoothing parameter
which is handed over to \code{\link{lowess}} as \code{f}}

\item{main}{\code{character} graphical parameter}

\item{xlab}{\code{character} graphical parameter}

\item{ylim}{(default if \code{rel = TRUE}: \code{c(0, 1)}) graphical parameter}

\item{both.lwd}{graphical parameter for smoothed values if \code{curves = "both"}}

\item{both.col}{graphical parameter for smoothed values if \code{curves = "both"}}

\item{both.lty}{graphical parameter for smoothed values if \code{curves = "both"}}

\item{natozero}{\code{logical} (default: \code{TRUE}) should NAs be coerced
to zeros. Only has effect if \code{rel = TRUE}.}

\item{file}{\code{character} file path if a pdf should be created}

\item{...}{additional graphical parameters}
}
\value{
A plot.
Invisible: A dataframe with columns \code{date} and \code{counts},
respectively \code{proportion}
}
\description{
Creates a plot of the counts/proportion of documents/words in the subcorpus,
which could be specified by \code{id}.
}
\details{
\code{object} needs a textmeta object with strictly tokenized text component
(character vectors) if you use \code{type = "words"}.
If you use \code{type = "docs"} use can use a tokenized or a not tokenized text component.
}
\examples{
##
}
\keyword{~kwd1}
\keyword{~kwd2}
