% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopic.R
\name{plotTopic}
\alias{plotTopic}
\title{Plotting Counts of Topics over Time (relative to Corpus)}
\usage{
plotTopic(object, ldaresult, ldaID, select = 1:nrow(ldaresult$document_sums),
  tnames, rel = FALSE, mark = TRUE, unit = "month", curves = c("exact",
  "smooth", "both"), smooth = 0.05, main, xlab, ylim, ylab, both.lwd,
  both.lty, col, legend = ifelse(pages, "onlyLast:topright", "topright"),
  pages = FALSE, natozero = TRUE, file, ...)
}
\arguments{
\item{object}{\code{\link{textmeta}} object with strictly tokenized
\code{text} component (\code{character} vectors) - like a result of
\code{\link{makeClear}}}

\item{ldaresult}{the result of a function call \code{\link{LDAstandard}}}

\item{ldaID}{\code{character} vector of IDs of the documents in
\code{ldaresult}}

\item{select}{\code{integer} vector (default: all topics) which topics of
\code{ldaresult} should be plotted}

\item{tnames}{\code{character} vector of same length as \code{select}
- labels for the topics (default are the first returned words of
\code{\link{top.topic.words}} from the \code{lda} package for each topic)}

\item{rel}{\code{logical} (default: \code{FALSE}) should counts
(\code{FALSE}) or proportion (\code{TRUE}) be plotted}

\item{mark}{\code{logical} (default: \code{TRUE}) should years be marked by
vertical lines}

\item{unit}{\code{character} (default: \code{"month"}) to which unit should
dates be floored. Other possible units are \code{"bimonth"}, \code{"quarter"}, \code{"season"},
\code{"halfyear"}, \code{"year"}, for more units see \code{\link[lubridate]{round_date}}}

\item{curves}{\code{character} (default: \code{"exact"}) should \code{"exact"},
\code{"smooth"} curve or \code{"both"} be plotted}

\item{smooth}{\code{numeric} (default: \code{0.05}) smoothing parameter
which is handed over to \code{\link{lowess}} as \code{f}}

\item{main}{\code{character} graphical parameter}

\item{xlab}{\code{character} graphical parameter}

\item{ylim}{graphical parameter}

\item{ylab}{\code{character} graphical parameter}

\item{both.lwd}{graphical parameter for smoothed values
if \code{curves = "both"}}

\item{both.lty}{graphical parameter for smoothed values
if \code{curves = "both"}}

\item{col}{graphical parameter, could be a vector. If \code{curves = "both"}
the function will for every topicgroup plot at first the exact and then the
smoothed curve - this is important for your col order.}

\item{legend}{\code{character} (default: \code{"topright"},
\code{"onlyLast:topright"} for \code{pages = TRUE} respectively)
value(s) to specify the legend coordinates. If "none" no legend is plotted.}

\item{pages}{\code{logical} (default: \code{FALSE}) should all curves be
plotted in a single plot. In addtion you could set
\code{legend = "onlyLast:<argument>"} with \code{<argument>} as a
\code{character} \code{legend} argument
for only plotting a legend on the last plot of set.}

\item{natozero}{\code{logical} (default: \code{TRUE}) should NAs be coerced
to zeros. Only has effect if \code{rel = TRUE}.}

\item{file}{\code{character} file path if a pdf should be created}

\item{...}{additional graphical parameters}
}
\value{
A plot.
Invisible: A dataframe with columns \code{date} and \code{tnames} with the
counts/proportion of the selected topics.
}
\description{
Creates a plot of the counts/proportion of specified topics of a result of
\code{\link{LDAstandard}}. There is an option to plot all curves in one plot
or to create one plot for every curve (see \code{pages}).
In addition the plots can be written to a pdf by setting \code{file}.
}
