% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWordpt.R
\name{plotWordpt}
\alias{plotWordpt}
\title{Plotting Counts of Topics-Words-Combination over Time (relative to Topics)}
\usage{
plotWordpt(object, docs, ldaresult, ldaID,
  select = 1:nrow(ldaresult$document_sums), link = c("and", "or"),
  wordlist = lda::top.topic.words(ldaresult$topics, 1), tnames, wnames,
  rel = FALSE, mark = TRUE, unit = "month", curves = c("exact",
  "smooth", "both"), smooth = 0.05, legend = ifelse(pages,
  "onlyLast:topright", "topright"), pages = FALSE, natozero = TRUE, file,
  main, xlab, ylab, ylim, both.lwd, both.lty, col, ...)
}
\arguments{
\item{object}{\code{\link{textmeta}} object with strictly tokenized
\code{text} component (\code{character} vectors) - like a result of
\code{\link{makeClear}}}

\item{docs}{object as a result of \code{\link{docLDA}} which was handed over
to \code{\link{LDAstandard}}}

\item{ldaresult}{the result of a function call \code{\link{LDAstandard}} with
\code{docs} as argument}

\item{ldaID}{\code{character} vector of IDs of the documents in
\code{ldaresult}}

\item{select}{list of \code{integer} vectors. Every list element is an 'or'
link, every \code{integer} string in a vector is linked by the argument
\code{link}. If \code{select} is only a \code{integer} vector it will be
coerced to a list of the same length as the vector (see \code{\link{as.list}}),
so that the argument \code{link} has no effect. Each \code{integer} vector
as a list element represents one curve in the outcoming plot}

\item{link}{\code{character} (default: \code{"and"}) should the (inner)
\code{integer} vectors of each list element be linked by an \code{"and"}
or an \code{"or"}}

\item{wordlist}{list of \code{character} vectors (default: the first
\code{\link{top.topic.words}} per topic as simple
\code{character} vector) which words - linked by an "or" every time -
should be take into account for plotting the topic counts/proportion.}

\item{tnames}{\code{character} vector of same length as \code{select}
- labels for the topics (default are the first returned words of}

\item{wnames}{\code{character} vector of same length as \code{wordlist}
- labels for every group of 'and' linked words
\code{\link{top.topic.words}} from the \code{lda} package for each topic)}

\item{rel}{\code{logical} (default: \code{FALSE}) should counts
(\code{FALSE}) or proportion (\code{TRUE}) be plotted}

\item{mark}{\code{logical} (default: \code{TRUE}) should years be marked by
vertical lines}

\item{unit}{\code{character} (default: \code{"month"}) to which unit should
dates be floored}

\item{curves}{\code{character} (default: \code{"exact"}) should \code{"exact"},
\code{"smooth"} curve or \code{"both"} be plotted}

\item{smooth}{\code{numeric} (default: \code{0.05}) smoothing parameter
which is handed over to \code{\link{lowess}} as \code{f}}

\item{legend}{\code{character} (default: \code{"topright"},
\code{"onlyLast:topright"} for \code{pages = TRUE} respectively)
value(s) to specify the legend coordinates. If "none" no legend is plotted.}

\item{pages}{\code{logical} (default: \code{FALSE}) should all curves be
plotted in a single plot. In addtion you could set
\code{legend = "onlyLast:<argument>"} with \code{<argument>} as a
\code{character} \code{legend} argument
for only plotting a legend on the last plot of set.}

\item{natozero}{\code{logical} (default: \code{TRUE}) should NAs be coerced
to zeros}

\item{file}{\code{character} file path if a pdf should be created}

\item{main}{\code{character} graphical parameter}

\item{xlab}{\code{character} graphical parameter}

\item{ylab}{\code{character} graphical parameter}

\item{ylim}{graphical parameter}

\item{both.lwd}{graphical parameter for smoothed values
if \code{curves = "both"}}

\item{both.lty}{graphical parameter for smoothed values
if \code{curves = "both"}}

\item{col}{graphical parameter, could be a vector. If \code{curves = "both"}
the function will for every wordgroup plot at first the exact and then the 
smoothed curve - this is important for your col order.}

\item{\dots}{additional graphical parameters}
}
\value{
A plot.
Invisible: A dataframe with columns \code{date} and \code{tnames: wnames}
with the counts/proportion of the selected combination of topics and words.
}
\description{
Creates a plot of the counts/proportion of specified combination of topics
and words. It is important to keep in mind that the baseline for
proportions are the sums of topics, not sums of words.
See also \code{\link{plotTopicWord}}.
There is an option to plot all curves in one plot or to create one plot for
every curve (see \code{pages}). In addition the plots can be written to a pdf
by setting \code{file}.
}
\examples{
##---- Should be DIRECTLY executable !! ----
}
\keyword{~kwd1}
\keyword{~kwd2}
