% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWordSub.R
\name{plotWordSub}
\alias{plotWordSub}
\title{Plotting Counts/Proportion of Words/Docs in LDA-generated Topic-Subcorpora over Time}
\usage{
plotWordSub(object, ldaresult, ldaID, limit = 10, alloc = c("multi",
  "unique", "best"), select = 1:nrow(ldaresult$document_sums), tnames, search,
  ignore.case = TRUE, type = c("docs", "words"), rel = TRUE,
  mark = TRUE, unit = "month", curves = c("exact", "smooth", "both"),
  smooth = 0.05, main, xlab, ylab, ylim, both.lwd, both.lty, col,
  legend = "topright", natozero = TRUE, file, ...)
}
\arguments{
\item{object}{\code{\link{textmeta}} object with strictly tokenized
\code{text} component (\code{character} vectors) - like a result of
\code{\link{makeClear}}}

\item{ldaresult}{the result of a function call \code{\link{LDAgen}}}

\item{ldaID}{\code{character} vector of IDs of the documents in
\code{ldaresult}}

\item{limit}{\code{integer/numeric} (default: \code{10}) the limit for how
often a word must be allocated to a topic to count these article as belonging
to this topic - if \code{0<limit<1} proportion is used}

\item{alloc}{\code{character} (default: \code{"multi"}) should every article
be allocated to multiple topics (\code{"multi"}), or maximum one
topic (\code{"unique"}), or the - exactly one - most represantative
topic (\code{"best"}). If \code{alloc = "best"} \code{limit} has no effect}

\item{select}{\code{integer} vector (default: all topics) which topics of
\code{ldaresult} should be plotted}

\item{tnames}{\code{character} vector of same length as \code{select}
- labels for the topics (default are the first returned words of
\code{\link{top.topic.words}} from the \code{lda} package for each topic)}

\item{search}{see \code{\link{filterWord}}}

\item{ignore.case}{see \code{\link{filterWord}}}

\item{type}{\code{character} (default: \code{"docs"}) should counts/proportion
of documents, where every  \code{"docs"} or words \code{"words"} be plotted}

\item{rel}{\code{logical} (default: \code{TRUE}) should counts
(\code{FALSE}) or proportion (\code{TRUE}) be plotted}

\item{mark}{\code{logical} (default: \code{TRUE}) should years be marked by
vertical lines}

\item{unit}{\code{character} (default: \code{"month"}) to which unit should
dates be floored}

\item{curves}{\code{character} (default: \code{"exact"}) should \code{"exact"},
\code{"smooth"} curve or \code{"both"} be plotted}

\item{smooth}{\code{numeric} (default: \code{0.05}) smoothing parameter
which is handed over to \code{\link{lowess}} as \code{f}}

\item{main}{\code{character} graphical parameter}

\item{xlab}{\code{character} graphical parameter}

\item{ylab}{\code{character} graphical parameter}

\item{ylim}{(default if \code{rel = TRUE}: \code{c(0, 1)}) graphical parameter}

\item{both.lwd}{graphical parameter for smoothed values
if \code{curves = "both"}}

\item{both.lty}{graphical parameter for smoothed values
if \code{curves = "both"}}

\item{col}{graphical parameter, could be a vector. If \code{curves = "both"}
the function will for every wordgroup plot at first the exact and then the
smoothed curve - this is important for your col order.}

\item{legend}{\code{character} (default: "topright") value(s) to specify the
legend coordinates. If "none" no legend is plotted.}

\item{natozero}{\code{logical} (default: \code{TRUE}) should NAs be coerced
to zeros. Only has effect if \code{rel = TRUE}.}

\item{file}{\code{character} file path if a pdf should be created}

\item{...}{additional graphical parameters}
}
\value{
A plot.
Invisible: A dataframe with columns \code{date} and \code{tnames} with the
counts/proportion of the selected topics.
}
\description{
Creates a plot of the counts/proportion of words/docs in corpora which are
generated by a \code{ldaresult}. Therefore an article is allocated to a topic
- and then to the topics corpus - if there are enough (see \code{limit} and
\code{alloc}) allocations of words in the article to the corresponding topic.
Additionally the corpora are reduced by \code{\link{filterWord}} and an
\code{search}-argument. The plot shows counts of subcorpora or if
\code{rel = TRUE} proportion of subcorpora on its corresponding whole corpora.
}
