% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.textmeta.R
\name{print.textmeta}
\alias{print.textmeta}
\title{Prints Objects of class "textmeta".}
\usage{
\method{print}{textmeta}(x, ...)
}
\arguments{
\item{x}{A list of corpus data of class "textmeta".}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
The dataset itself.
}
\description{
Prints a given Dataset, which should be of class "textmeta".
}
\examples{

}
\author{
Jonas Rieger (<riegerjonas@gmx.de>)
}
\keyword{manip}

