% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.summary.R
\name{read.summary}
\alias{read.summary}
\title{Print Summarys of Datasets.}
\usage{
read.summary(data, list.names = names(data))
}
\arguments{
\item{data}{A list of corpus data as a result of a read-function.}

\item{list.names}{character: Which list.elements should be summarized?}
}
\value{
The Dataset itself.
}
\description{
Prints a summary of a given dataset, which should be a result of a read-funtion.
}
\examples{

}
\author{
Jonas Rieger (<riegerjonas@gmx.de>)
}
\keyword{manip}

