% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{read_survey}
\alias{read_survey}
\title{Read survey data into a DataFrame}
\usage{
read_survey(filename)
}
\arguments{
\item{filename}{Path to the CSV file to read}
}
\description{
Skip the second row which contains the survey questions
}
\details{
Thanks to https://stackoverflow.com/a/15860268/3420371
}
