% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readJFArchiv.R
\name{readJFArchiv}
\alias{readJFArchiv}
\title{Read the JF-Archiv.de Corpus as CSV}
\usage{
readJFArchiv(path = getwd(), file = list.files(path = path, pattern =
  "*.csv$", full.names = FALSE, recursive = TRUE), do.meta = TRUE,
  do.text = TRUE)
}
\arguments{
\item{path}{Character string with Path where the data files are.}

\item{file}{Character string with names of the CVS files.}

\item{do.meta}{Logical: Should the algorithm collect meta data?}

\item{do.text}{Logical: Should the algorithm collect text data?}
}
\value{
\item{meta}{ id date title year number page_start page_stop pagetitle
shorttitle rubrik ressort dokumentmerkmal dachzeile abstract}
\item{text}{ Text (Paragraphenweise)} \item{metamult}{ signature person
koerperschaft company inkl. Kategorie(n)}
}
\description{
Reads the CSV-files from a WORDPRESS BLOG corpus and seperates the text and meta
data.
}
\examples{
##---- Should be DIRECTLY executable !! ----
}
\keyword{manip}
