% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNexis.R
\name{readNexis}
\alias{readNexis}
\title{Read preprocessed files from Lexis Nexis}
\usage{
readNexis(file, do.meta = TRUE, do.text = TRUE, encoding = "utf-8")
}
\arguments{
\item{file}{Character string with names of the XML files.}

\item{do.meta}{Logical: Should the algorithm collect meta data?}

\item{do.text}{Logical: Should the algorithm collect text data?}

\item{encoding}{encoding of the input files.}
}
\value{
List of two \item{meta}{id url date page resource author leadtext downloadDate}
\item{text}{text}
}
\description{
Reads the XML-files from preprocessed Lexis Nexis files and seperates the text and meta
data.
}
\examples{

}
\author{
Sakander Zirai (<s.zirai@live.de>), Lars Koppers (<koppers@statistik.tu-dortmund.de>)
}
\keyword{manip}

