% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNexisOnline.R
\name{readNexisOnline}
\alias{readNexisOnline}
\title{Read preprocessed files from Nexis Online}
\usage{
readNexisOnline(path = getwd(), file = list.files(path = path, pattern =
  "*.HTML", full.names = FALSE, recursive = TRUE), do.meta = TRUE,
  do.text = TRUE)
}
\arguments{
\item{path}{Path where the data files are.}

\item{file}{Character string with names of the HTML files.}

\item{do.meta}{Logical: Should the algorithm collect meta data?}

\item{do.text}{Logical: Should the algorithm collect text data?}
}
\value{
List of two \item{meta}{id topic nr from title source date releaseNote
downloadDate loadDate language length dateline byline section type pubType series graphic}
\item{text}{text}
}
\description{
Reads the HTML-files from preprocessed Lexis Nexis files and seperates the
text and meta data.
}
\examples{

}
\author{
Jonas Rieger (<riegerjonas@gmx.de>)
}
\keyword{manip}

