% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSZ.R
\name{readSZ}
\alias{readSZ}
\title{Read the SZ corpus}
\usage{
readSZ(file, folder, do.meta = TRUE, do.text = TRUE)
}
\arguments{
\item{file}{Names of the XML files.}

\item{folder}{Names of the folder, in which the files are. Must have same
length like \code{file}.}

\item{do.meta}{Logical: Should the algorithm collect meta data?}

\item{do.text}{Logical: Should the algorithm collect text data?}
}
\value{
\item{meta}{ id datum rubrik seite AnzChar AnzWoerter dachzeile
titel utitel} \item{text}{ Text (Paragraphenweise)}
}
\description{
Reads the XML-files from the SZ corpus and seperates the text and meta data.
}
\examples{

##---- Should be DIRECTLY executable !! ----
}
\author{
Lars Koppers (<koppers@statistik.tu-dortmund.de>)
}
\keyword{manip}

