% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWhatsapp.R
\name{readWhatsApp}
\alias{readWhatsApp}
\title{Read WhatsApp files}
\usage{
readWhatsApp(path = getwd(), file = list.files(path = path, pattern =
  "*.html$", full.names = FALSE, recursive = TRUE))
}
\arguments{
\item{path}{Character: string with path where the data files are.}

\item{file}{Character: string with names of the HTML files.}

\item{encoding}{encoding of the input files.}
}
\value{
\code{\link{textmeta}} object.
}
\description{
Reads HTML-files from WhatsApp and separates the text and meta data.
}
\author{
Jonas Rieger (<jonas.rieger@tu-dortmund.de>)
}
\keyword{manip}
