% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWikinews.R
\name{readWikinews}
\alias{readWikinews}
\title{Read the SZ corpus}
\usage{
readWikinews(path = getwd(), file = list.files(path = path, pattern =
  "*.xml$", full.names = FALSE, recursive = TRUE))
}
\arguments{
\item{path}{Path where the data files are.}

\item{file}{Character string with names of the HTML files.}
}
\value{
textmeta-object
}
\description{
Reads the XML-files from the Wikinews Export page \url{https://en.wikinews.org/wiki/Special:Export}.
}
\examples{

##---- Should be DIRECTLY executable !! ----
}
