% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeMISC.R
\name{removeMISC}
\alias{removeMISC}
\title{Removes trailing and leading whitespaces and Original Gesamtseiten-PDF}
\usage{
removeMISC(text)
}
\arguments{
\item{text}{character string including XML-Tags.}
}
\description{
Removes trailing and leading whitespaces and Original Gesamtseiten-PDF
}
\examples{

}
\author{
Sakander Zirai (<s.zirai@live.de>)
}
\keyword{manip}

