% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeTAG.R
\name{removeTAG}
\alias{removeTAG}
\title{Removes XML Tags}
\usage{
removeTAG(text)
}
\arguments{
\item{text}{character string including XML-Tags.}
}
\description{
Removes XML Tags
}
\examples{

}
\author{
Sakander Zirai (<s.zirai@live.de>)
}
\keyword{manip}

