% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeXML.R
\name{removeXML}
\alias{removeXML}
\alias{removeUmlauts}
\alias{removeHTML}
\title{Removes XML tags and umlauts}
\usage{
removeXML(x)

removeUmlauts(x)

removeHTML(x, dec = TRUE, hex = TRUE, entity = TRUE, symbolList = 1,
  delete = TRUE, symbols = FALSE)
}
\arguments{
\item{x}{\code{character} vector.}

\item{dec}{logical remove html decimal?}

\item{hex}{logical remove html hex?}

\item{entity}{logical remove html name?}

\item{symbolList}{Which ISO-8859 Lists?}

\item{delete}{delete other entitys?}

\item{symbols}{check symbols?}

\item{xml}{\code{logical} Should XML-TAGs be removed?}

\item{umlauts}{\code{logical}: Should umlauts be changed?}

\item{u.type}{Type of umlaut changing: \code{normal} for normal umlauts 
-> ue), \code{html} for html representation of umlauts (&uuml; -> ue) and
\code{all} for both.}

\item{html}{Logical: Should Numeric character references beeing deleted (after replacing umlauts etc.)}
}
\value{
Adjusted corpus
}
\description{
Removes XML tags and changes umlauts to a standardized form.
}
\details{
The decision which u.type is used should consider the language of the corpus, because in some languages the replacement of umlauts can change the meaning of the a word.
To change which columns are used by removeXML use argument xmlAction in \code{\link{readTextmeta}}.
}
\examples{


##---- Should be DIRECTLY executable !! ----
}
\keyword{manip}
