% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.articles.R
\name{show.articles}
\alias{show.articles}
\title{Export Readable Meta-Data of Articles.}
\usage{
show.articles(meta, file, id = meta$id, cols = colnames(meta),
  csv.ger = FALSE)
}
\arguments{
\item{meta}{A data.frame of meta-data as a result of a read-function.}

\item{file}{Filename for the export.}

\item{id}{Vector of characters including article ids.}

\item{cols}{vector of characters including the requested colnames of meta.}

\item{csv.ger}{Logical: Should the function use write.csv2 instead of write.csv?}
}
\value{
No output in R, writes a csv (or csv2) including the meta-data of the
requested articles.
}
\description{
Exports requested meta-data of articles for given ids.
}
\examples{

}
\author{
Jonas Rieger (<riegerjonas@gmx.de>)
}
\keyword{manip}

