% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showArticles.R
\name{showArticles}
\alias{showArticles}
\title{Export Readable Article Lists}
\usage{
showArticles(object, id = names(object$text),
  file = deparse(substitute(object)))
}
\arguments{
\item{object}{\code{\link{textmeta}} object}

\item{id}{Vector or matrix inkluding article id's.}

\item{file}{Filename for the export.}
}
\value{
No output in R, only a csv including the requested articles.
}
\description{
Exports the article id, text, title and date.
}
\examples{

##---- Should be DIRECTLY executable !! ----
}
\keyword{manip}
