% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showMeta.R
\name{showMeta}
\alias{showMeta}
\title{Export Readable Meta-Data of Articles.}
\usage{
showMeta(meta, id = meta$id, cols = colnames(meta), file)
}
\arguments{
\item{meta}{A data.frame of meta-data as a result of a read-function.}

\item{id}{Character vector or matrix including article ids.}

\item{cols}{Character vector including the requested columns of meta.}

\item{file}{Character Filename for the export.}
}
\value{
A list of the requested meta data. If file is set, writes a csv including the meta-data of the
requested meta data.
}
\description{
Exports requested meta-data of articles for given ids.
}
\examples{

}
\keyword{manip}
