% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcorpus.date.R
\name{subcorpus.date}
\alias{subcorpus.date}
\title{Subcorpus With Date Filter}
\usage{
subcorpus.date(text, meta, s.date = min(meta$datum),
  e.date = max(meta$datum))
}
\arguments{
\item{text}{List of article texts}

\item{meta}{Dataframe of metadata}

\item{s.date}{Start date of subcorpus as date object}

\item{e.date}{End date of subcorpus as date object}
}
\value{
Filtered list of texts.
}
\description{
Creates a subcorpus including a specific time span.
}
\examples{

##---- Should be DIRECTLY executable !! ----
}
\author{
Lars Koppers (<koppers@statistik.tu-dortmund.de>)
}
\keyword{manip}
