% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcorpus.words.R
\name{subcorpus.words}
\alias{subcorpus.words}
\title{Subcorpus With Word Filter}
\usage{
subcorpus.words(text, wordlist, counts = 1, ignore.case = FALSE,
  perl = FALSE, fixed = FALSE, useBytes = FALSE)
}
\arguments{
\item{text}{List of article texts}

\item{wordlist}{List of character vectors. Every List element is an 'or'
link, every character String in a vector is linked by an 'and'. If
\code{wordlist} is only a character Vector the link is 'or'.}

\item{counts}{Integer, or same list structure like \code{wordlist}. Number
of times a word must appear to be counted.}

\item{ignore.case}{Option from \code{\link{grepl}}.}

\item{perl}{Option from \code{\link{grepl}}.}

\item{fixed}{Option from \code{\link{grepl}}.}

\item{useBytes}{Option from \code{\link{grepl}}.}
}
\value{
Filtered list of texts.
}
\description{
Creates a subcorpus including specific words.
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (text, wordlist, counts = 1, ignore.case = FALSE, perl = FALSE,
    fixed = FALSE, useBytes = FALSE)
{
    subid <- numeric(length(text))
    for (i in 1:length(wordlist)) {
        tmp <- NULL
        for (j in 1:length(wordlist[[i]])) {
            if (length(counts) == 1) {
                co <- counts
            }
            else {
                co <- counts[[i]][j]
            }
            tmp <- cbind(tmp, sapply(text, function(x) sum(grepl(pattern = wordlist[[i]][j],
                x = x, ignore.case = ignore.case, perl = perl,
                fixed = fixed, useBytes = useBytes)) >= co))
        }
        subid <- subid + apply(tmp, 1, prod)
    }
    subid <- subid > 0
    return(text[subid])
  }

}
\author{
Lars Koppers (<koppers@statistik.tu-dortmund.de>)%% ~~who you are~~
}
\keyword{~kwd1}
\keyword{~kwd2}

