% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcorpusCount.R
\name{subcorpusCount}
\alias{subcorpusCount}
\title{Subcorpus With Count Filter}
\usage{
subcorpusCount(object, text, count, out = c("text", "bin", "count"))
}
\arguments{
\item{object}{\code{\link{textmeta}} object}

\item{text}{not necassary if \code{object} is specified, else should be
\code{object\$text}: list of article texts}

\item{count}{an integer marking how many words must at least be found in the text.}

\item{out}{Type of output: \code{text} filtered corpus,
\code{bin} logical vector for all texts,
\code{count} the counts.}
}
\value{
Filtered list of texts.
}
\description{
Creates a subcorpus including specific articles.
}
\examples{

##---- Should be DIRECTLY executable !! ----
}
\keyword{~kwd1}
\keyword{~kwd2}
