% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcorpusDate.R
\name{subcorpusDate}
\alias{subcorpusDate}
\title{Subcorpus With Date Filter}
\usage{
subcorpusDate(text, meta, s.date = min(meta$date), e.date = max(meta$date))
}
\arguments{
\item{text}{List of article texts}

\item{meta}{Dataframe of metadata}

\item{s.date}{Start date of subcorpus as date object}

\item{e.date}{End date of subcorpus as date object}
}
\value{
Filtered list of texts.
}
\description{
Creates a subcorpus including a specific time span.
}
\examples{

##---- Should be DIRECTLY executable !! ----
}
\keyword{manip}

