% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.textmeta.R
\name{summary.textmeta}
\alias{summary.textmeta}
\title{Print Summarys of Objects of class "textmeta".}
\usage{
\method{summary}{textmeta}(object, list.names = names(object), ...)
}
\arguments{
\item{object}{A list of corpus data of class "textmeta".}

\item{list.names}{character: Which list.elements should be summarized?}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
The dataset itself.
}
\description{
Prints a summary of a given Dataset, which should be of class "textmeta".
}
\examples{

}
\author{
Jonas Rieger (<riegerjonas@gmx.de>)
}
\keyword{manip}

