% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.textmeta.R
\name{tidy.textmeta}
\alias{tidy.textmeta}
\title{Transform textmeta to an object with tidy text data}
\usage{
tidy.textmeta(object)
}
\arguments{
\item{object}{A \code{\link{textmeta}} object}
}
\value{
An object with tidy text data
}
\description{
Transfers data from a text component of a \code{\link{textmeta}} object to a
tidy data.frame.
}
\keyword{manip}
