\name{tmt-package}
\alias{tmt-package}
\alias{tmt}
\docType{package}
\title{
Text Manipulation Toolkit
}
\description{
This is a package for cleaning, reformating, and otherwise manipulating text data.
}
\details{
\tabular{ll}{
Package: \tab tmt\cr
Type: \tab Package\cr
Version: \tab 0.01\cr
Date: \tab 2013-14-02\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}

}
\author{
Maintainer: Schaun Wheeler [schaun.wheeler@gmail.com)
}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~
%~~ directory ~~
\keyword{
	tmt
	text mining
	text analysis
}
\seealso{
	\code{\link[tmt:AspellCheck]{AspellCheck}}
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
%\examples{
%  rga.open(instance="ga")
%}
