% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmt.R
\docType{package}
\name{tmt}
\alias{tmt}
\alias{tmt-package}
\title{tmt: A package for loading survey data and generating reports}
\description{
The tmt package provides two categories of important functions:
load and report.
}
\section{Load functions}{

}

\section{Report functions}{

}

