% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmt_interpret.R
\name{tmt_interpret}
\alias{tmt_interpret}
\title{interprets the TMT results}
\usage{
tmt_interpret(test = NULL, norm = "tombaugh", score = NULL)
}
\arguments{
\item{test}{character. the name of the test. this name is only
used in the shiny app for interpreting the results}

\item{norm}{the normative data. currently only \code{"tombaugh"} is implemented}

\item{score}{the score of the test}
}
\description{
interprets the TMT results
}
