% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tombaugh.norm.R
\name{tombaugh.norm}
\alias{tombaugh.norm}
\title{tombaugh.norm}
\usage{
tombaugh.norm(age = NULL, education = NULL, tmtA = NULL, tmtB = NULL)
}
\description{
the normative data are taken from: Tombaugh, T. N. (2004). Trail Making Test A and B: normative data stratified by age and education. Archives of Clinical Neuropsychology, 19(2), 203–214.
}
