% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicOverTime.R
\name{topicOverTime}
\alias{topicOverTime}
\title{Plotting Topics Over Time}
\usage{
topicOverTime(x, ldaID, meta = NULL, file, topicNo = NULL, Tnames = NULL,
  unit = "month", ...)
}
\arguments{
\item{x}{LDA result object}

\item{ldaID}{Character vector including IDs of the texts.}

\item{meta}{The meta data for the texts or a date-string.}

\item{file}{Name of the pdf file.}

\item{topicNo}{character string of used topicnumbers}

\item{Tnames}{Label for the topics}

\item{unit}{unit for \code{\link{round_date}}}

\item{\dots}{Further Statements passed to the plot function.}
}
\value{
A pdf.
}
\description{
Creates a pdf including a plot for each topic. For each topic the number of
words per month would be plotted.
}
\examples{

##---- Should be DIRECTLY executable !! ----
}
\keyword{~kwd1}
\keyword{~kwd2}
