% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicwords.over.time.R
\name{topicwords.over.time}
\alias{topicwords.over.time}
\title{Plotting Topicwords Over Time}
\usage{
topicwords.over.time(lda.corpus, meta, lda.vocab, lda.result, woerter, topics,
  file, lda.name)
}
\arguments{
\item{lda.corpus}{Corpus of type lda.}

\item{meta}{The meta data for the texts}

\item{lda.vocab}{Character vector of the lda vocabulary}

\item{lda.result}{LDA result object}

\item{woerter}{List of word vectors.}

\item{topics}{Same structure like \code{woerter}. Each Entry gives the
number of the topic for the word at the same position in \code{woerter}.}

\item{file}{Name of the pdf file}

\item{lda.name}{Name of the lda, as a lable for the pdf output.}
}
\value{
A pdf.
}
\description{
Creates a pdf including a plot for each choosen word. For each word the
number of this word per month in the whole corpus and in the choosen topic
would be plotted.
}
\examples{

##---- Should be DIRECTLY executable !! ----
}
\author{
Lars Koppers (<koppers@statistik.tu-dortmund.de>)
}
\keyword{~kwd1}
\keyword{~kwd2}

