% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicwordsOverTime.R
\name{topicwordsOverTime}
\alias{topicwordsOverTime}
\title{Plotting Topicwords Over Time}
\usage{
topicwordsOverTime(ldaCorpus, meta, ldaVocab, ldaResult, words, topics, file,
  ldaName, unit = "month", ...)
}
\arguments{
\item{ldaCorpus}{Corpus of type lda.}

\item{meta}{The meta data for the texts}

\item{ldaVocab}{Character vector of the lda vocabulary}

\item{ldaResult}{LDA result object}

\item{topics}{Same structure like \code{woerter}. Each Entry gives the
number of the topic for the word at the same position in \code{woerter}.}

\item{file}{Name of the pdf file}

\item{ldaName}{Name of the lda, as a label for the pdf output.}

\item{unit}{unit for\link{\code{round_date}}}

\item{\dots}{additional parameter for \link{\code{plot}}}

\item{woerter}{List of word vectors.}
}
\value{
A pdf.
}
\description{
Creates a pdf including a plot for each choosen word. For each word the
number of this word per month in the whole corpus and in the choosen topic
would be plotted.
}
\examples{

##---- Should be DIRECTLY executable !! ----
}
\keyword{~kwd1}
\keyword{~kwd2}
