% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tot.heat.R
\name{tot.heat}
\alias{tot.heat}
\title{Plotting Topics over Time relative to Corpus}
\usage{
tot.heat(x, topics = 1:nrow(x$document_sums), ldaID, meta = NULL,
  corpus = NULL, norm = FALSE, file, Tnames = top.topic.words(x$topics,
  1), date_breaks = "1 year")
}
\arguments{
\item{x}{LDA result object.}

\item{topics}{Numeric vector containing the numbers of the topics to be plotted. Defaults to all topics.}

\item{ldaID}{Character vector containing IDs of the texts.}

\item{meta}{Optional. Specify to analyze subcorpus. The meta data for the texts. One of meta or corpus has to be specified.}

\item{corpus}{Optional. Specify to analyze entire copous. The data used for normalization. One of meta or corpus has to be specified.}

\item{norm}{Logical. Should the values be normalized by the mean topic share to account for differently sized topics? Defaults to FALSE.}

\item{file}{Character vector containing the path and name for the pdf output file.}

\item{Tnames}{Character vector with labels for the topics.}

\item{date_breaks}{Which years should be shown on the x axis. Can be one of "1 year","5 years" or "10 years".}
}
\value{
A pdf.
}
\description{
Creates a pdf including a heat map. For each topic, the heat map shows the deviation of 
its current share from its mean share. Shares can be calculated on subcorpus or corpus level.
Shares can be calculated in absolute deviation from the mean or relative to the mean of the topic to account for different topic strengths.
}
\examples{
##
}
\author{
Kira Schacht (<kira.schacht@tu-dortmund.de>)
}
\keyword{~kwd1}
\keyword{~kwd2}
