% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tot.relative.R
\name{tot.relative}
\alias{tot.relative}
\title{Plotting Topics over Time relative to Corpus}
\usage{
tot.relative(topics = 1:nrow(x$document_sums), x, ldaID, meta = NULL,
  corpus = NULL, file, pages = TRUE, Tnames = top.topic.words(x$topics,
  1), smooth = 0.05)
}
\arguments{
\item{topics}{Numbers of the topics to be plotted. Defaults to all topics.}

\item{x}{LDA result object.}

\item{ldaID}{Character vector including IDs of the texts.}

\item{meta}{Specify to analyze subcorpus. The meta data for the texts.}

\item{corpus}{Specify to analyze entiere corous. The data used for normalization.}

\item{file}{Name of the pdf file.}

\item{pages}{Should the topics be plotted on separate pages (true) or on one page (false). Defaults to true.}

\item{Tnames}{Vector with labels for the topics.}

\item{smooth}{How much the output should be smoothed. Set to 0 to disable smoothing.}
}
\value{
A pdf.
}
\description{
Creates a pdf including a plot for each topic. For each topic the share of
words per month would be plotted. Shares can be calculated on subcorpus or corpus level.
}
\examples{
##
}
\author{
Kira Schacht (<kira.schacht@tu-dortmund.de>)
}
\keyword{~kwd1}
\keyword{~kwd2}

