% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTopics.R
\name{clusterTopics}
\alias{clusterTopics}
\title{Cluster Analysis}
\usage{
clusterTopics(topics, file, topicnames = NULL, method = "average",
  width = 30, height = 15, ...)
}
\arguments{
\item{topics}{A matrix like the \code{topic} matrix from
\code{lda.collapse.gibbs.sampler}}

\item{file}{File for the dendogram pdf.}

\item{topicnames}{Character vector as label for the topics.}

\item{method}{Method statement from \code{\link{hclust}}}

\item{width}{grafical parameter for pdf output. See \code{\link{pdf}}}

\item{height}{grafical parameter for pdf output. See \code{\link{pdf}}}

\item{...}{additional parameter for \code{\link{plot}}}
}
\value{
A dendogram as pdf and a List \item{dist}{A distance matrix}
\item{clust}{The result from \code{hclust}.}
}
\description{
Makes a cluster analysis with the hellinger distance.
}
\examples{

##---- Should be DIRECTLY executable !! ----
}
\keyword{~kwd1}
\keyword{~kwd2}
