% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteAndRenameDuplicates.R
\name{deleteAndRenameDuplicates}
\alias{deleteAndRenameDuplicates}
\title{Delete And Rename Articles with the same ID}
\usage{
deleteAndRenameDuplicates(object, paragraph = FALSE)
}
\arguments{
\item{object}{A textmeta-object as a result of a read-function.}

\item{paragraph}{Logical: Should be set to \code{TRUE} if the article is a
list of character strings, representing the paragraphs.}
}
\description{
Deletes Articles with the same ID and same text. Renames the ID of Articles
with the same ID but different text-component (_IDFakeDup, _IDRealDup).
}
