% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intruderTopics.R
\name{intruderTopics}
\alias{intruderTopics}
\title{Function to fit LDA model}
\usage{
intruderTopics(text = NULL, beta = NULL, theta = NULL, id = NULL,
  numIntruder = 1, numOuttopics = 4, byScore = TRUE, minWords = 0L,
  minOuttopics = 0L, stopTopics = NULL, printSolution = FALSE,
  oldResult = NULL, test = FALSE, testinput = NULL)
}
\arguments{
\item{text}{tba}

\item{beta}{A matrix of word-probabilities or frequency table for the topics. Each row is a topic, each column a word. The rows will be divided by the row sums, if they are not 1.}

\item{theta}{tba}

\item{id}{tba}

\item{numIntruder}{intended number of intruder words. If \code{numIntruder} is a integer vector, the number would be sampled for each topic.}

\item{numOuttopics}{tba integer. Number of words per topic, including the intruder words.}

\item{byScore}{logical should the score of \code{top.topic.words} from the \code{lda} package be used?}

\item{minWords}{tba}

\item{minOuttopics}{tba}

\item{stopTopics}{tba}

\item{printSolution}{tba}

\item{oldResult}{result object from an unfinished run of \code{intruderWords}. If oldResult is used, all other parameter will be ignored.}

\item{test}{logical enables test mode}

\item{testinput}{input for function tests}
}
\value{
tba
}
\description{
tba
}
\examples{

##---- Should be DIRECTLY executable !! ----
}
\references{
tba
}
\seealso{
tba
}
\author{
Lars Koppers (<koppers@statistik.tu-dortmund.de>)
}
\keyword{~kwd1}
\keyword{~kwd2}
