% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAstandard.R
\name{LDAstandard}
\alias{LDAstandard}
\title{Function to fit LDA model}
\usage{
LDAstandard(documents, K = 100L, vocab, num.iterations = 200L,
  burnin = 70L, alpha = NULL, eta = NULL, seed, folder, num.words = 50L,
  LDA = TRUE)
}
\arguments{
\item{documents}{A list prepared by \code{\link{docLDA}}.}

\item{K}{number of topics.}

\item{vocab}{character vector containing the words in the corpus.}

\item{num.iterations}{number of iterations for the gibbs sampler.}

\item{burnin}{number of iterations for the burnin.}

\item{alpha}{Hyperparameter for the topic proportions.}

\item{eta}{Hyperparameter for the word distributions.}

\item{seed}{A seed for reproducability.}

\item{folder}{file for the results.}

\item{num.words}{number of words in the top topic words list.}

\item{LDA}{logical: Should a new model be fitted or a existing R workspace
be used?}
}
\value{
A .csv containing the topword list and a R workspace containing the
result data.
}
\description{
This function uses the \code{lda.collapsed.gibbs.sampler} from the LDA
package and additional saves topwordlists and a R workspace.
}
\examples{

##---- Should be DIRECTLY executable !! ----
}
\references{
Blei, David M. and Ng, Andrew and Jordan, Michael. Latent
Dirichlet allocation. Journal of Machine Learning Research, 2003.

Jonathan Chang (2012). lda: Collapsed Gibbs sampling methods for topic
models.. R package version 1.3.2. http://CRAN.R-project.org/package=lda
}
\seealso{
Documentation for the lda package.
}
\keyword{~kwd1}
\keyword{~kwd2}

