% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.wordlist.R
\name{make.wordlist}
\alias{make.wordlist}
\title{Count words in text corpora}
\usage{
make.wordlist(x, k = 1e+05)
}
\arguments{
\item{x}{List of texts.}

\item{k}{Integer, how many Texts should be processed at one time (RAM
usage)?}
}
\value{
\item{words}{An alphabetical list of the words in the corpus.}
\item{wordtable}{A frequency table of the words in the corpus.}
}
\description{
Creates a wordlist and a frequency table.
}
\details{
This function helps, if \code{table(x)} needs too much RAM.
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x, k = 1e+05)
{
    n <- length(x)
    (print(n))
    N <- 0:floor(n/k)
    words <- NULL
    for (i in N) {
        (print(i * k))
        words <- c(words, unique(unlist(x[(i * k + 1):(min(n,
            i * k + k))])))
    }
    words <- sort(unique(words))
    (print("table"))
    wordtable <- rep(0, length(words))
    names(wordtable) <- words
    for (i in N) {
        (print(i * k))
        tmp <- table(unlist(x[(i * k + 1):(min(n, i * k + k))]))
        mtch <- match(names(tmp), names(wordtable))
        wordtable[mtch] <- wordtable[mtch] + tmp
    }
    return(list(words = words, wordtable = wordtable))
  }

}
\author{
Lars Koppers (<koppers@statistik.tu-dortmund.de>)
}
\keyword{manip}

