% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeClear.R
\name{makeClear}
\alias{makeClear}
\title{Some data preprocessing}
\usage{
makeClear(x, sw = c(stopwords("german"), "dass", "fuer", "koennen", "koennte",
  "ueber", "waehrend", "wuerde", "wuerden"), paragraph = TRUE)
}
\arguments{
\item{x}{list containing character strings.}

\item{sw}{list of stopwords.}

\item{paragraph}{Logical: Should be set to \code{TRUE} if the article is a
list of character strings, representing the paragraphs.}
}
\value{
A list containing the preprocessed Article.
}
\description{
Removes punctuation, numbers and stopwords, change into lowercase letters
and tokenization.
}
\details{
Removes punctuation, numbers and stopwords, change into lowercase letters
and tokenization. Additional some cleaning steps: remove empty words /
paragraphs / article.
}
\examples{

##---- Should be DIRECTLY executable !! ----
}
\keyword{manip}
