% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWordlist.R
\name{makeWordlist}
\alias{makeWordlist}
\title{Count words in text corpora}
\usage{
makeWordlist(text, k = 100000L)
}
\arguments{
\item{text}{List of texts.}

\item{k}{Integer, how many Texts should be processed at one time (RAM
usage)?}
}
\value{
\item{words}{An alphabetical list of the words in the corpus.}
\item{wordtable}{A frequency table of the words in the corpus.}
}
\description{
Creates a wordlist and a frequency table.
}
\details{
This function helps, if \code{table(x)} needs too much RAM.
}
\examples{

##---- Should be DIRECTLY executable !! ----
}
\keyword{manip}
