% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.cluster.R
\name{match.cluster}
\alias{match.cluster}
\title{Preparation of different LDAs For Clustering}
\usage{
match.cluster(x)
}
\arguments{
\item{x}{A list of lda results.}
}
\value{
A matrix including all topics from all lda-results. Number of rows
is the number of topics, the number of columns is the number of words which
appear in all results.
}
\description{
Merges different lda-results to one matrix, including only the words which
appears in all lda.
}
\examples{

##---- Should be DIRECTLY executable !! ----
}
\author{
Lars Koppers (<koppers@statistik.tu-dortmund.de>)
}
\keyword{manip}

