% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeTextmeta.R
\name{mergeTextmeta}
\alias{mergeTextmeta}
\title{Merge textmeta objects}
\usage{
mergeTextmeta(x, all = TRUE)
}
\arguments{
\item{x}{a list of \code{\link{textmeta}} objects}

\item{all}{\code{logical} (default: \code{TRUE}) should the result contain
\code{\link{union}} (\code{TRUE}) or \code{\link{intersect}} of columns of
all objects, if \code{TRUE} columns which only appears in one of
the meta components the merged meta component is filled with \code{NA}s}
}
\value{
\code{\link{textmeta}} object
}
\description{
Merges a list of textmeta objects to one. It is possible to control whether all
columns or the intersect should be considered.
}
\examples{

}
\keyword{manip}
