% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScot.R
\name{plotScot}
\alias{plotScot}
\title{Plotting Counts of Documents or Words over Time (relative to Corpus)}
\usage{
plotScot(object, ids = object$meta$id, type = c("docs", "words"),
  rel = FALSE, mark = TRUE, main, xlab, ylim, ...)
}
\arguments{
\item{object}{\code{\link{textmeta}} object - with strictly tokenized
\code{text} component (\code{character} vectors) if \code{type = "words"}}

\item{ids}{\code{character} vector (default: \code{object$meta$id}) which IDs
specify the subcorpus}

\item{type}{\code{character} (default: \code{"docs"}) should counts/proportion
of documents \code{"docs"} or words \code{"words"} be plotted}

\item{rel}{\code{logical} (default: \code{FALSE}) should counts
(\code{FALSE}) or proportion (\code{TRUE}) be plotted}

\item{mark}{\code{logical} (default: \code{TRUE}) should years be marked by
vertical lines}

\item{main}{\code{character} graphical parameter}

\item{xlab}{\code{character} graphical parameter}

\item{ylim}{(default if \code{rel = TRUE}: \code{c(0, 1)}) graphical parameter}

\item{...}{additional graphical parameters}
}
\value{
A plot.
Invisible: A dataframe with \item{date} and \item{counts} or \item{proportion}.
}
\description{
Creates a plot of the counts/proportion of documents/words in the subcorpus,
which could be specified by \code{ids}.
}
