% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.summary.R
\name{read.summary}
\alias{read.summary}
\title{Print Summarys of Datsets.}
\usage{
read.summary(data, metamult = "metamult" \%in\% colnames(data))
}
\arguments{
\item{data}{A list of corpus data as a result of a read-function.}

\item{metamult}{Logical: Should the function return a summary of data$metamult?}
}
\value{
A print in the console.
}
\description{
Prints a summary of a given dataset, which should be a result of a read-funtion.
}
\examples{

}
\author{
Jonas Rieger (<riegerjonas@gmx.de>)
}
\keyword{manip}

