% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHBWiWo.R
\name{readHBWiWo}
\alias{readHBWiWo}
\title{Read the HB WiWo Corpus}
\usage{
readHBWiWo(file, en, do.meta = TRUE, do.text = TRUE, test = FALSE)
}
\arguments{
\item{file}{Names of the XML files.}

\item{en}{Logical: Are there english descriptions in the XML structure? Must
have same length like \code{file}.}

\item{do.meta}{Logical: Should the algorithm collect meta data?}

\item{do.text}{Logical: Should the algorithm collect text data?}

\item{test}{logical. other directory}
}
\value{
\item{meta}{ id quelle datum titel abstract UB dachzeile}
\item{text}{ Text} \item{metamult}{ person firma industrie land autor rubrik
klassifikation (mehrere moeglich) thema sachgruppe serie}
}
\description{
Reads the XML-files from the HB WiWo corpus and seperates the text and meta
data.
}
\examples{

}
\author{
Lars Koppers (<koppers@statistik.tu-dortmund.de>)
}
\keyword{manip}

