% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHBWiWo.R
\name{readHBWiWo}
\alias{readHBWiWo}
\title{Read the HB WiWo Corpus}
\usage{
readHBWiWo(path = getwd(), file = list.files(path = path, pattern =
  "*.xml$", full.names = FALSE, recursive = TRUE), do.meta = TRUE,
  do.text = TRUE)
}
\arguments{
\item{path}{Character string with Path where the data files are.}

\item{file}{Character string with names of the XML files.}

\item{do.meta}{Logical: Should the algorithm collect meta data?}

\item{do.text}{Logical: Should the algorithm collect text data?}
}
\value{
\item{meta}{ id source date title abstract dachzeile}
\item{text}{ Text} \item{metamult}{ person company industry country author category
klassifikation (mehrere moeglich) thema sachgruppe serie}
}
\description{
Reads the XML-files from the HB WiWo corpus and seperates the text and meta
data.
}
\keyword{manip}
