% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSPIEGEL.R
\name{readSPIEGEL}
\alias{readSPIEGEL}
\title{Read the SPIEGEL Corpus}
\usage{
readSPIEGEL(path = getwd(), file = list.files(path = path, pattern =
  "*.xml", full.names = F, recursive = TRUE), do.meta = TRUE,
  do.text = TRUE)
}
\arguments{
\item{path}{Character string with Path where the data files are.}

\item{file}{Character string with names of the XML files.}

\item{do.meta}{Logical: Should the algorithm collect meta data?}

\item{do.text}{Logical: Should the algorithm collect text data?}
}
\value{
\item{meta}{ id jahr nummer datum seiteS seiteE seitentitel
kurztitel rubrik ressort dokumentenmerkmal dachzeile titel vorspann}
\item{text}{ Text (Paragraphenweise)} \item{metamult}{ person koerperschaft
firma inkl. Kategorie(n)}
}
\description{
Reads the XML-files from the SPIEGEL corpus and seperates the text and meta
data.
}
\examples{
##---- Should be DIRECTLY executable !! ----
}
\keyword{manip}

