% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeXML.R
\name{removeXML}
\alias{removeXML}
\title{Removes XML tags and umlauts}
\usage{
removeXML(x, xml = TRUE, umlauts = FALSE, u.type = c("normal", "html",
  "all"), remove.html = TRUE)
}
\arguments{
\item{x}{\code{character} vector.}

\item{xml}{\code{logical} Should XML-TAGs be removed?}

\item{umlauts}{\code{logical}: Should umlauts be changed?}

\item{u.type}{Type of umlaut changing: \code{normal} for normal umlauts (\enc{ü}{u}
-> ue), \code{html} for html representation of umlauts (&uuml; -> ue) and
\code{all} for both.}

\item{remove.html}{Logical: Should Numeric character references beeing deleted (after replacing umlauts etc.)}
}
\value{
Adjusted corpus
}
\description{
Removes XML tags and changes umlauts to a standardized form.
}
\examples{


##---- Should be DIRECTLY executable !! ----
}
\keyword{manip}
