% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeXML.R
\name{removeXML}
\alias{removeXML}
\alias{removeUmlauts}
\alias{removeHTML}
\title{Removes XML Tags and Umlauts}
\usage{
removeXML(x)

removeUmlauts(x)

removeHTML(x, dec = TRUE, hex = TRUE, entity = TRUE, symbolList = c(1:4,
  9, 13, 15, 16), delete = TRUE, symbols = FALSE)
}
\arguments{
\item{x}{Character: Vector}

\item{dec}{logical remove html decimal?}

\item{hex}{logical remove html hex?}

\item{entity}{logical remove html name?}

\item{symbolList}{Which ISO-8859 Lists?}

\item{delete}{delete other entitys?}

\item{symbols}{check symbols?}
}
\value{
Adjusted corpus
}
\description{
Removes XML tags and changes umlauts in a standardized form.
}
\details{
The decision which u.type is used should consider the language of the corpus, because in some languages the replacement of umlauts can change the meaning of a word.
To change which columns are used by removeXML use argument xmlAction in \code{\link{readTextmeta}}.
}
\examples{
xml <- "<text>Some <b>important</b> text</text>"
removeXML(xml)

x <- "&#x00f8; &#248; &oslash;"
removeHTML(x=x, symbolList = 1, dec=TRUE, hex=FALSE, entity=FALSE, delete = FALSE)
removeHTML(x=x, symbolList = c(1,3))

y <- c("Bl\\UFChende Apfelb\\UE4ume")
removeUmlauts(y)

}
\keyword{manip}
