% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeXML.R
\name{removeXML}
\alias{removeXML}
\title{Removes XML tags and umlauts}
\usage{
removeXML(x, xml = TRUE, umlauts = FALSE, u.type = c("normal", "html",
  "all"))
}
\arguments{
\item{x}{List of texts.}

\item{xml}{Logical: Should XML-TAGs be removed?}

\item{u.type}{Type of umlaut changing: \code{normal} for normal umlauts (\enc{ü}{u}
-> ue), \code{html} for html representation of umlauts (&uuml; -> ue) and
\code{all} for both.}

\item{umlaute}{Logical: Should umlauts be changed?}
}
\value{
Adjusted corpus
}
\description{
Removes XML tags and changes umlauts to a standardized form.
}
\examples{


##---- Should be DIRECTLY executable !! ----
}
\author{
Lars Koppers (<koppers@statistik.tu-dortmund.de>)
}
\keyword{manip}

