% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showArticles.R
\name{showArticles}
\alias{showArticles}
\title{Export Readable Article Lists}
\usage{
showArticles(object, id = names(object$text), file)
}
\arguments{
\item{object}{\code{\link{textmeta}} object}

\item{id}{Character vector or matrix including article ids.}

\item{file}{Character Filename for the export.}
}
\value{
A list of the requested articles. If file is set, writes a csv including the meta-data of the
requested articles.
}
\description{
Exports the article id, text, title and date.
}
\examples{

##---- Should be DIRECTLY executable !! ----
}
\keyword{manip}
