% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcorpus.words.R
\name{subcorpus.words}
\alias{subcorpus.words}
\title{Subcorpus With Word Filter}
\usage{
subcorpus.words(text, wordlist, counts = 1, ignore.case = FALSE,
  perl = FALSE, fixed = FALSE, useBytes = FALSE, out = c("text", "bin",
  "count"), ...)
}
\arguments{
\item{text}{List of article texts}

\item{wordlist}{List of character vectors. Every List element is an 'or'
link, every character String in a vector is linked by an 'and'. If
\code{wordlist} is only a character Vector the link is 'or'.}

\item{counts}{Integer, or same list structure like \code{wordlist}. Number
of times a word must appear to be counted.}

\item{ignore.case}{Option from \code{\link{grepl}}.}

\item{perl}{Option from \code{\link{grepl}}.}

\item{fixed}{Option from \code{\link{grepl}}.}

\item{useBytes}{Option from \code{\link{grepl}}.}

\item{out}{Type of output: \code{text} filtered corpus, \code{bin} logical vector for all texts, \code{count} the number of matches (max one match per character string).}

\item{...}{additional parameters for \code{grepl}}
}
\value{
Filtered list of texts.
}
\description{
Creates a subcorpus including specific words.
}
\examples{

##---- Should be DIRECTLY executable !! ----
}
\author{
Lars Koppers (<koppers@statistik.tu-dortmund.de>)%% ~~who you are~~
}
\keyword{~kwd1}
\keyword{~kwd2}

