% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcorpusDate.R
\name{subcorpusDate}
\alias{subcorpusDate}
\title{Subcorpus With Date Filter}
\usage{
subcorpusDate(object, text = object$text, meta = object$meta,
  s.date = min(meta$date, na.rm = TRUE), e.date = max(meta$date, na.rm =
  TRUE))
}
\arguments{
\item{object}{\code{\link{textmeta}} object}

\item{text}{not necassary if \code{object} is specified, else should be
\code{object\$text}}

\item{meta}{not necassary if \code{object} is specified, else should be
\code{object\$meta}}

\item{s.date}{Start date of subcorpus as date object}

\item{e.date}{End date of subcorpus as date object}
}
\value{
\code{\link{textmeta}} object if \code{object} is specified,
else only the filtered \code{text}.
}
\description{
Creates a subcorpus including a specific time span.
}
\examples{

##---- Should be DIRECTLY executable !! ----
}
\keyword{manip}
