% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.textmeta.R
\name{summary.textmeta}
\alias{summary.textmeta}
\title{Print Summarys of Objects of class "textmeta".}
\usage{
\method{summary}{textmeta}(data, list.names = names(data))
}
\arguments{
\item{data}{A list of corpus data of class "textmeta".}

\item{list.names}{character: Which list.elements should be summarized?}
}
\value{
The Dataset itself.
}
\description{
Prints a summary of a given dataset, which should be of class "textmeta".
}
\examples{

}
\author{
Jonas Rieger (<riegerjonas@gmx.de>)
}
\keyword{manip}

