% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmeta.R
\name{textmeta}
\alias{textmeta}
\alias{is.textmeta}
\alias{print.textmeta}
\alias{summary.textmeta}
\title{"textmeta"-Objects}
\usage{
textmeta(meta = NULL, text = NULL, metamult = NULL,
  dateFormat = "\%Y-\%m-\%d")

is.textmeta(x)

\method{print}{textmeta}(x, ...)

\method{summary}{textmeta}(object, listnames = names(object),
  metavariables = character(), ...)
}
\arguments{
\item{meta}{data.frame (or matrix) of the meta-data.}

\item{text}{list (or character vector) of the text-data.}

\item{metamult}{list of the metamult-data.}

\item{dateFormat}{\code{character} string with the date format in meta.
for \code{\link{as.Date}}}

\item{x}{an R Object.}

\item{...}{further arguments in print and summary. Not implemented.}

\item{object}{textmeta object}

\item{listnames}{Character vector with names of textmeta lists (meta, text, metamult). Summaries are generated for those lists only. Default gives summaries for all lists.}

\item{metavariables}{Character vector with variable-names from the meta dataset. Summaries are generated for those variables only.}
}
\value{
a "textmeta"-Object.
}
\description{
Creates "textmeta"-Objects.

Tests for "textmeta"-Objects.
}
\examples{

}
\keyword{manip}
