% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmeta.R
\name{textmeta}
\alias{textmeta}
\title{"textmeta"-Objects}
\usage{
textmeta(meta = NULL, text = NULL, metamult = NULL)
}
\arguments{
\item{meta}{data.frame of the meta-data.}

\item{text}{list or character of the text-data.}

\item{metamult}{list of the metamult-data.}

\item{x}{an R Object.}
}
\value{
\item{textmeta}{ returns a "textmeta"-Object.}
\item{is.textmeta}{ returns TRUE or FALSE.}
}
\description{
Creates or tests for "textmeta"-Objects.
}
\examples{

}
\author{
Jonas Rieger (<riegerjonas@gmx.de>)
}
\keyword{manip}

