\name{tmT-package}
\alias{tmT-package}
\alias{tmT}
\docType{package}
\title{
\packageTitle{tmT}
}
\description{
\packageDescription{tmT}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tmT}
\packageIndices{tmT}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{tmT}

Maintainer: \packageMaintainer{tmT}
}
%% \references{
%% ~~ Literature or other references for background information ~~
%% }
%% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%% ~~ the R documentation directory ~~
\keyword{ package }
%% \seealso{
%% ~~ Optional links to other man pages, e.g. ~~
%% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%% }
%% \examples{
%% ~~ simple examples of the most important functions ~~
%% }
