% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmT-package.R
\docType{package}
\name{tmT-package}
\alias{tmT}
\alias{tmT-package}
\title{c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_title(\"#1\")}",
"tmT")\Sexpr{tools:::Rd_package_title("tmT")}}
\description{
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_description(\"#1\")}",
"tmT")\Sexpr{tools:::Rd_package_description("tmT")}
}
\details{
The DESCRIPTION file:
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_DESCRIPTION(\"#1\")}",
"tmT")\Sexpr{tools:::Rd_package_DESCRIPTION("tmT")}
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_indices(\"#1\")}",
"tmT")\Sexpr{tools:::Rd_package_indices("tmT")} ~~ An overview of how to use
the package, including the most important ~~ ~~ functions ~~
}
\author{
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_author(\"#1\")}",
"tmT")\Sexpr{tools:::Rd_package_author("tmT")}

Maintainer:
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_maintainer(\"#1\")}",
"tmT")\Sexpr{tools:::Rd_package_maintainer("tmT")}
}
\keyword{package}

