% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toparticles.R
\name{toparticles}
\alias{toparticles}
\title{Get The IDs Of The Most Representive Articles}
\usage{
toparticles(x, id, limit = 20, rel = TRUE, topn = TRUE,
  themes = 1:nrow(x$document_sums), minlength = 30)
}
\arguments{
\item{x}{LDA result}

\item{id}{Vector of text IDs}

\item{limit}{Integer, number of article ids per topic.}

\item{rel}{Logical: Should be the relative frequency be used?}

\item{topn}{Logical: Should only the topn Article IDs be reported?
\code{FALSE} is not implemented.}

\item{themes}{not implemented.}

\item{minlength}{Minimal total number of words a text must have to be
included.}
}
\value{
Matrix of Article IDs.
}
\description{
Extract the article IDs belonging to the articles with the most relative or
absolute number of words per topic.
}
\examples{

##---- Should be DIRECTLY executable !! ----
}
\author{
Lars Koppers (<koppers@statistik.tu-dortmund.de>)
}
\keyword{manip}

