% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topArticles.R
\name{topArticles}
\alias{topArticles}
\title{Get The IDs Of The Most Representive Articles}
\usage{
topArticles(ldaresult, ldaID, limit = 20L, rel = TRUE,
  select = 1:nrow(ldaresult$document_sums), tnames, minlength = 30L)
}
\arguments{
\item{ldaresult}{LDA result}

\item{ldaID}{Vector of text IDs}

\item{limit}{Integer, number of article ids per topic.}

\item{rel}{Logical: Should be the relative frequency be used?}

\item{select}{which topics should be returned?}

\item{tnames}{names of the selected topics}

\item{minlength}{Minimal total number of words a text must have to be
included.}
}
\value{
Matrix of Article IDs.
}
\description{
Extract the article IDs belonging to the articles with the most relative or
absolute number of words per topic.
}
\examples{

##---- Should be DIRECTLY executable !! ----
}
\keyword{manip}
