% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topArticles.R
\name{topArticles}
\alias{topArticles}
\title{Get The IDs Of The Most Representive Articles}
\usage{
topArticles(x, id, limit = 20L, rel = TRUE, themes = NULL,
  minlength = 30L)
}
\arguments{
\item{x}{LDA result}

\item{id}{Vector of text IDs}

\item{limit}{Integer, number of article ids per topic.}

\item{rel}{Logical: Should be the relative frequency be used?}

\item{themes}{which topics should be returned?}

\item{minlength}{Minimal total number of words a text must have to be
included.}
}
\value{
Matrix of Article IDs.
}
\description{
Extract the article IDs belonging to the articles with the most relative or
absolute number of words per topic.
}
\examples{

##---- Should be DIRECTLY executable !! ----
}
\keyword{manip}
