% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic.cluster.R
\name{topic.cluster}
\alias{topic.cluster}
\title{Cluster Analysis}
\usage{
topic.cluster(topics, file, topicnames = 1:nrow(topics), method = "average")
}
\arguments{
\item{topics}{A matrix like the \code{topic} matrix from
\code{lda.collapse.gibbs.sampler}}

\item{file}{File for the dendogram pdf.}

\item{topicnames}{Character vector as label for the topics.}

\item{method}{Method statement from \code{\link{hclust}}}
}
\value{
A dendogram as pdf and a List \item{dist}{A distance matrix}
\item{clust}{The result from \code{hclust}.}
}
\description{
Makes a cluster analysis with the hellinger distance.
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (topics, file, topicnames = 1:nrow(topics), method = "average")
{
    topics <- topics/rowSums(topics)
    topics <- sqrt(topics)
    Dist <- 1/sqrt(2) * dist(topics)
    clust <- hclust(d = Dist, method)
    pdf(file, width = 30, height = 15)
    plot(clust, label = topicnames)
    dev.off()
    return(list(Dist, clust))
  }

}
\author{
Lars Koppers (<koppers@statistik.tu-dortmund.de>)
}
\keyword{~kwd1}
\keyword{~kwd2}

