% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic.over.time.R
\name{topic.over.time}
\alias{topic.over.time}
\title{Plotting Topics Over Time}
\usage{
topic.over.time(x, ldaID, meta, file, Tnames, ...)
}
\arguments{
\item{x}{LDA result object}

\item{ldaID}{Character vector including IDs of the texts.}

\item{meta}{The meta data for the texts.}

\item{file}{Name of the pdf file.}

\item{Tnames}{Label for the topics}

\item{\dots}{Further Statements passed to the plot function.}
}
\value{
A pdf.
}
\description{
Creates a pdf including a plot for each topic. For each topic the number of
words per month would be plotted.
}
\examples{

##---- Should be DIRECTLY executable !! ----
}
\author{
Lars Koppers (<koppers@statistik.tu-dortmund.de>)
}
\keyword{~kwd1}
\keyword{~kwd2}

