% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tot.relative.R
\name{tot.relative}
\alias{tot.relative}
\title{Plotting Topics over Time relative to Corpus}
\usage{
tot.relative(x, topics = 1:nrow(x$document_sums), ldaID, meta = NULL,
  corpus = NULL, file, pages = TRUE, Tnames = top.topic.words(x$topics,
  1), smooth = 0.05)
}
\arguments{
\item{x}{LDA result object.}

\item{topics}{Numeric vector containing the numbers of the topics to be plotted. Defaults to all topics.}

\item{ldaID}{Character vector containing IDs of the texts.}

\item{meta}{Optional. Specify to analyze subcorpus. The meta data for the texts. One of meta or corpus has to be specified.}

\item{corpus}{Optional. Specify to analyze entire copous. The data used for normalization. One of meta or corpus has to be specified.}

\item{file}{Character vector containing the path and name for the pdf output file.}

\item{pages}{Logical. Should the topics be plotted on separate pages (true) or on one page (false). Defaults to true.}

\item{Tnames}{Character vector with labels for the topics.}

\item{smooth}{How much the output should be smoothed. The higher the number, the more smoothing is performed. Set to 0 to disable smoothing.}
}
\value{
A pdf.
}
\description{
Creates a pdf including a plot for each topic. For each topic the share of
words per month would be plotted. Shares can be calculated on subcorpus or corpus level.
}
\examples{
##
}
\author{
Kira Schacht (<kira.schacht@tu-dortmund.de>)
}
\keyword{~kwd1}
\keyword{~kwd2}

